keeping up

+++ app/src/main/java/com/duckduckgo/app/bookmarks/db/BookmarkEntity.kt
@@ -16,10 +16,12 @@
 
 package com.duckduckgo.app.bookmarks.db
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity(tableName = "bookmarks")
-data class BookmarkEntity(@PrimaryKey(autoGenerate = true) var id: Int = 0,
-                          var title: String?,
-                          var url: String)
+data class BookmarkEntity(
+    @PrimaryKey(autoGenerate = true) var id: Long = 0,
+    var title: String?,
+    var url: String
+)
